requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary
pymongo
python-dotenv